fixed bugs with token, removed incorrect indexSelector, transfered logic from recipes to users regarding favorites
