+++ src/components/Banner/Banner.module.css
@@ -13,9 +13,9 @@
   margin-left: calc(-50vw + 50%);
 }
 
-
 /* Retina Mobile */
-@media screen and (min-resolution: 192dpi), screen and (min-device-pixel-ratio: 2) {
+@media screen and (min-resolution: 192dpi),
+  screen and (min-device-pixel-ratio: 2) {
   .bannerSection {
     background-image: url("/images/BannerMob@2x.png");
   }
@@ -42,7 +42,6 @@
   background: linear-gradient(to right, rgba(0, 0, 0, 0.9), rgba(0, 0, 0, 0.3));
 }
 
-
 /*   &::before {
     content: "";
     position: absolute;
@@ -55,9 +54,9 @@
       to right,
       rgba(0, 0, 0, 0.9),
       rgba(0, 0, 0, 0.3)
-    ); */
+    ); 
   }
-}
+} */
 
 .boxTitle {
   position: absolute;
@@ -93,7 +92,8 @@
     background-image: url("/images/Tablet@1x.png");
   }
 
-  @media screen and (min-resolution: 192dpi), screen and (min-device-pixel-ratio: 2) {
+  @media screen and (min-resolution: 192dpi),
+    screen and (min-device-pixel-ratio: 2) {
     .bannerSection {
       background-image: url("/images/Tablet@2x.png");
     }
@@ -117,7 +117,8 @@
     background-image: url("/images/Desctop.png");
   }
 
-  @media screen and (min-resolution: 192dpi), screen and (min-device-pixel-ratio: 2) {
+  @media screen and (min-resolution: 192dpi),
+    screen and (min-device-pixel-ratio: 2) {
     .bannerSection {
       background-image: url("/images/Desctop@2x.png");
     }
@@ -131,5 +132,4 @@
   .titleBanner {
     font-size: 48px;
   }
-
 }
