+++ src/redux/operations/recipesOperation.js
@@ -54,33 +54,6 @@ export const fetchRecipeByIdThunk = createAsyncThunk(
   }
 );
 
-export const fetchFavoriteRecipesThunk = createAsyncThunk(
-  "fetchFavoriteRecipes",
-  async (_, thunkAPI) => {
-    try {
-      const { data } = await API.get("recipes/favorites");
-      return data;
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error.message);
-    }
-  }
-);
-
-// {action} means ADD or REMOVE = like a toggle
-// backend can find it in req.body
-
-export const updateFavoriteRecipesThunk = createAsyncThunk(
-  "updateFavoriteRecipes",
-  async ({ id, action }, thunkAPI) => {
-    try {
-      await API.patch("recipes/favorites", { recipeId: id, action });
-      return { id, action };
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error.message);
-    }
-  }
-);
-
 export const fetchOwnRecipesThunk = createAsyncThunk(
   "fetchOwnRecipes",
   async (_, thunkAPI) => {
@@ -92,6 +65,7 @@ export const fetchOwnRecipesThunk = createAsyncThunk(
     }
   }
 );
+
 export const searchRecipesThunk = createAsyncThunk(
   "recipes/search",
   async (
@@ -114,4 +88,3 @@ export const searchRecipesThunk = createAsyncThunk(
     }
   }
 );
-
