+++ src/redux/operations/userOperation.js
@@ -12,3 +12,33 @@ export const fetchUserThunk = createAsyncThunk(
     }
   }
 );
+
+export const fetchFavoriteRecipesThunk = createAsyncThunk(
+  "fetchFavoriteRecipes",
+  async (_, thunkAPI) => {
+    try {
+      const { data } = await API.get("users/favorites");
+      return data;
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error.message);
+    }
+  }
+);
+
+// {action} means ADD or REMOVE = like a toggle
+// backend can find it in req.body
+
+export const updateFavoriteRecipesThunk = createAsyncThunk(
+  "updateFavoriteRecipes",
+  async ({ id, action }, thunkAPI) => {
+    try {
+      const { data } = await API.patch("users/favorites", {
+        recipeId: id,
+        action,
+      });
+      return data;
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error.message);
+    }
+  }
+);
