+++ src/redux/selectors/authSelector.js
@@ -1,6 +1,5 @@
-export const selectAuthIsLoggedIn = (state) => state.auth.selectAuthIsLoggedIn;
+export const selectAuthIsLoggedIn = (state) => state.auth.isLoggedIn;
 export const selectAuthIsRefreshing = (state) => state.auth.isRefreshing;
 export const selectAuthIsLoading = (state) => state.auth.isLoading;
 export const selectAuthError = (state) => state.auth.error;
 export const selectAuthToken = (state) => state.auth.token;
-export const selectAuthUser = (state) => state.auth.user;
