--- src/redux/selectors/index.js
@@ -1,5 +1,0 @@
-export * from "./authSelectors";
-export * from "./recipesSelectors";
-export * from "./categoriesSelectors";
-export * from "./ingredientsSelectors";
-export * from "./userSelectors";
