+++ src/redux/selectors/recipesSelector.js
@@ -3,5 +3,4 @@ export const selectRecipesError = (state) => state.recipes.error;
 export const selectRecipes = (state) => state.recipes.items;
 export const selectRecipeDetails = (state) => state.recipes.recipeDetails;
 export const selectFilters = (state) => state.recipes.filters;
-export const selectFavoriteRecipes = (state) => state.recipes.favoriteRecipes;
 export const selectOwnRecipes = (state) => state.recipes.ownRecipes;
