+++ src/redux/slices/userSlice.js
@@ -1,5 +1,9 @@
-import { createSlice } from "@reduxjs/toolkit";
-import { fetchUserThunk } from "../operations/userOperation";
+import { createSlice, isAnyOf } from "@reduxjs/toolkit";
+import {
+  fetchFavoriteRecipesThunk,
+  fetchUserThunk,
+  updateFavoriteRecipesThunk,
+} from "../operations/userOperation";
 
 const initialState = {
   user: {
@@ -29,7 +33,38 @@ const userReducer = createSlice({
       .addCase(fetchUserThunk.rejected, (state, { payload }) => {
         state.isLoading = false;
         state.error = payload;
-      }),
+      })
+      .addMatcher(
+        isAnyOf(
+          fetchFavoriteRecipesThunk.fulfilled,
+          updateFavoriteRecipesThunk.fulfilled
+        ),
+        (state, { payload }) => {
+          state.user.favorites = payload;
+          state.isLoading = false;
+          state.error = false;
+        }
+      )
+      .addMatcher(
+        isAnyOf(
+          fetchFavoriteRecipesThunk.pending,
+          updateFavoriteRecipesThunk.pending
+        ),
+        (state) => {
+          state.isLoading = true;
+          state.error = false;
+        }
+      )
+      .addMatcher(
+        isAnyOf(
+          fetchFavoriteRecipesThunk.rejected,
+          updateFavoriteRecipesThunk.rejected
+        ),
+        (state, { payload }) => {
+          state.isLoading = false;
+          state.error = payload;
+        }
+      ),
 });
 
 export default userReducer.reducer;
